Fixed bugs and added component #309
Fixed a bug where it would only mention one dependency at a time, added a component to avoid adding too much complexity to TableListView.tsx
